refactor: unify template metadata hooks
